// src/app/models/contactdata.ts
export class Contact {
    id: number;
    firstName: string;
    lastName: string;
    position: string;
    address: string;
    imageUrl: string;
    companyName: string;
    companyAddrss: string;
    companyCity: string;
    companyState: string;
    companyZip: string;
    companyPhone: string;
}